+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added `GetAccountProofs` endpoint (#506).
 - Migrated faucet from actix-web to axum (#511).
 - Changed the `BlockWitness` to pass the inputs to the VM using only advice provider (#516).
+- [BREAKING] Improved store API errors (return "not found" instead of "internal error" status if requested account(s) not found) (#518).
 
 ## 0.5.1 (2024-09-12)